Initialize MockTexeraDB using JDBC (#1919)
This PR addresses the issue of frequent test failures that started
occurring after commit #1912. The failure is caused by the
initialization of the mock Texera DB. Upon running the Texera DDL
script, the DB exits due to a known issue in the MariaDB4j library, as
documented in this GitHub issue:
https://github.com/MariaDB4j/MariaDB4j/issues/233.

To resolve this problem, I have identified several potential solutions:

1. Upgrade the library to version 3.0.0. However, this requires Java 17,
and migrating our codebase to Java 17 would involve a substantial
effort.
2. Downgrade the library to version 2.3.0. However, many of the
functionalities we rely on are not yet available in this version.
3. Upon further investigation, I discovered that the error comes from a
managed process in MariaDB4j which executes the SQL. To solve this
issue, we can execute the SQL script through JDBC, thereby avoiding the
usage of the managed process.