Fix Property Editor Text Field Uneditable On Creation (#1921)
This PR fixes an annoying bug described as below:

In frontend's property editor, for dynamically created properties (e.g.,
UDF's new attributes, Projection fields, etc.), when a new property
subfield of an `Array` type property is created (e.g., a UDF's new
attribute), if the subfield contains a textual field, the text editor
will not be editable. Clicking on the blank canvas and clicking back
onto the operator will make the text field editable.

This issue is only observable on production server. The reason of this
bug is because of improper type equivalence checking code in shared
property editor. The old code checks if the newly created field is a
sub-field of an array (because it does not know during run-time, since
this is a custom formly field wrapper) by comparing the parent object's
constructor name with `YArray`. This works locally because the
constructor name for a `YArray` is `YArray` as-is. However, on
production, the object type is encapsulated and renamed as `r6`, which
make this piece of code invalid.

The fix is to simply change this type checking part to using
`instanceof`, which is a proper way to do this. I tested and this works
both locally and on production.