Refactor UserWorkflowComponent into SortButtonComponent, FiltersComponent (#1953)
1. Refactored UserWorkflowComponent into SortButtonComponent,
FiltersComponent.
2. Updated UserWorkflowService to maintain a list of projects. When user
changes the color of a project, the service notifies all subscribers of
the color change. This allow the new change to be reflected in the UI in
UserWorkflowListItemComponent. Previously, it was done with a
hack through two way bindings of a variable which triggers an update API
call in UserWorkflowComponent.
3. The FiltersComponent now contain all the logics of writing filters
into the search box. It outputs the masterFilterList to be consumed
by UserWorkflowCompoennt, which makes the API call to search and display
results.
4. Crerated DashboardEntry which is a view model over
DashboardWorkflowEntry, UserProject, DashboardFile
5. Moved existing unit tests into the corresponding components.