Remove frontend timeout for execution/pause/resume requests (#2004)
Currently, if our backend takes longer than 3 seconds to respond, the
frontend says the request failed, even when it didn't. This could
confuse users and cause them to start another execution unnecessarily.
The plan is to move this limit to the backend instead. This PR removes
timeout in the frontend code.