Replace out-of-date library json-ref-lite (#1788)
The library [json-ref-lite](https://www.npmjs.com/package/json-ref-lite)
is outdated and lacks maintenance. It depends on an old version of
property-expr@^1.3.1, which has a [critical
vulnerability](https://github.com/Texera/texera/security/dependabot/133).

This PR replaces `json-ref-lite` with
[@stoplight/json-ref-resolver](https://www.npmjs.com/package/@stoplight/json-ref-resolver)
to remove the dependency of `property-expr`.

As of 06/25/2023,
[@stoplight/json-ref-resolver](https://www.npmjs.com/package/@stoplight/json-ref-resolver)
has been also deprecated in favor of
[json-schema-ref-parser](https://github.com/APIDevTools/json-schema-ref-parser).
However, the latter is also in maintenance mode and waiting for new
maintainers. Will consider other packages in the future.