Fix python worker proxy server failed to start issue (#2030)
In this PR, we have made some enhancements to the process of starting
the server and client. Originally, the port number for python server is
determined by the Java side, which will cause the problem that the
python server might fail to start. The Java side gets the free port
number by starting a socket and releasing it, but this cannot guarantee
that the port is still available when python receives it. Now, the Java
side no longer determines the port number for the python side. Instead,
the Python server now selects an available port directly, which can
ensure its availability. The updated whole process of starting server
and client is presented below:
![2 drawio
(5)](https://github.com/Texera/texera/assets/114192306/37dd6cb9-4bb7-4672-b3af-16cd52eaf62f)

In conclusion, this PR contains:
1. Optimize the server-client process, and let servers determine their
port number by themselves.
2. Make sure the server start correctly. It will try again and again
until it succeeds.
Noticed that there are still other potential issues such as related to
running the second time with large number of workers is not included in
this PR.

Here it shows that 100 workers can work:
![Peek 2023-06-27
12-04](https://github.com/Texera/texera/assets/114192306/2f045f22-3bae-4fd3-af02-043b455b7f06)

---------

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>