Adding maximize and minimize button for html/url visualizer (#2070)
In this PR, we introduced the functionality of two dynamic resizing
buttons for the HTML visualizer. The two buttons switch between each
other based on the window's current size. A maximize button is displayed
when the window is at its original dimensions, and the users can expand
the visualizer size to the entire window. Conversely, when the window is
in full-window mode, a minimize button is displayed and the users can
revert the visualizer back to its original size.

Here is a demo of this functionality:
![Peek 2023-07-18
23-59](https://github.com/Texera/texera/assets/114192306/24d806f5-de9d-4198-b7a2-17f23a3b379e)