Refactor Python Table and TableLike (#2067)
This PR refactors the `Table` object and `TableLike` in the Python UDF.

## TableLike
`TableLike` is now defined as a TypeVar, which can be viewed as a
collection of `Tuple` instances. The following types are considered as
valid `TableLike`:
- Table
- pandas.DataFrame
- List[TupleLike]
  - List[Tuple]
  - List[Dict[str, Field]]

## Table

To support easier usage, `Table` still inherits `pandas.DataFrame`.
We provide multiple conversion methods to help convert other `TableLike`
instances into a Table. We also add converters between `TableLike` and
`Table` to ease the UDF code writing. With those converters, this PR
also fixes an issue #2068 that was caused by previous improper
conversions from a `Table` to a collection of `Tuple`s.