Add TreeMap Visualization Operator (#2075)
This PR introduces a new visualization operator: TreeMap

Treemaps display hierarchical data as a set of nested rectangles. Each
branch of the tree is given a rectangle, which is then tiled with
smaller rectangles representing sub-branches. A leaf node's rectangle
has an area proportional to a specified dimension of the data. Often the
leaf nodes are colored to show a separate dimension of the data.

The user needs to specify a value column for the data associated with
each node in the tree and also provides an ordered array of columns to
specify a path from the root to the leaves.

The operator filters missing values and 0s from the input table
automatically to try its best to make successful visualizations.

Example of visualization:
![____2023-07-21 __1 03
20](https://github.com/Texera/texera/assets/13672781/5685d8f5-8195-4a0f-95ff-53156d7c859d)

---------

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>