Fix outer join in HashJoinOp (#2110)
This PR fixes a bug on HashJoinOp, where it complains about schema
mismatch during an outer join. This happens when the
`buildAttributeName` is different from the `probeAttributeName`. The
DESC logic will use the `buildAttribtueName` in the final outer join
result tuple. However, the EXEC expects the `probeAttributeName`
instead. This PR aligns them to use the `buildAttributeName ` in both
DESC and EXEC.