Add a hyperlink to the Texera main logo on the navigation page for system users (#2132)
This PR adds a hyperlink to the Texera logo on the navigation page for
system users. The page will redirect to the dashboard if a system user
clicks the logo. The behavior is the same when clicking the dashboard
button on the navigation page. If the user is not a system user, the
hyperlink will be disabled.

[System User]
![Texera - Chrome 2023-08-31
14-28-02_edited_trimmed](https://github.com/Texera/texera/assets/143021053/77cb255b-331d-4547-abc1-6004b7787c81)

[Non-system User]
![Texera - Chrome 2023-08-31
14-39-02](https://github.com/Texera/texera/assets/143021053/37270aa0-acc0-442d-a1e8-4cffab392012)

---------

Co-authored-by: Xiaozhen Liu <xiaozl3@uci.edu>