Add DumbbellPlot Viz Operator (#2149)
This pull request introduces the DumbbellPlot viz operator.

A Dumbbell Plot, often referred to as a Dumbbell Chart, is a type of
data visualization primarily used to compare the difference between two
set of data points. This operator allows user to select which two sets
of data to compare, which measurement to compare, and entities involved
in the comparison.

For implementation details, this operator is a python-native operator.
It accepts the user-specified data and generate corresponding python
script. The figure is created with `plotly.graph_objects.Figure`. More
details about the plotly Dumbbell chart drawing is available
[here](https://plotly.com/python/dumbbell-plots/).

Here is a small demo:
![2023-09-19 17 35
56](https://github.com/Texera/texera/assets/43344272/e2861808-c919-4761-99ab-1046d6b38831)


The icon image is from <a
href="https://www.freepik.com/icon/dumbbell_206254">Icon by Freepik</a>