Bug Fix for Pagination Control Disappearing Issue (Issue #2147) (#2153)
This PR is a bug fix of [this
issue](https://github.com/Texera/texera/issues/2147).

The main problem is when all execution entries on the current page are
deleted, the pagination control disappears, making it impossible for the
user to view other entries without reopening the window. This issue
arises when the current page has no data to display. However, there is
an existing function called changePaginatedExecutions() that can update
the execution entries on the current page. I have reused this code to
consistently update the paginated execution entries after deletion. This
means that other execution results will fill the page after all
execution results on the current page are deleted.

However, there is a special scenario to consider when deleting all
execution entries on the last page. When this happens, the last page
won't have any data to display which causes the same issue. To address
this, I have added a special logic so that after deletion, if there are
no execution entries on the current page after update, it will
automatically move to the previous page index.


![capture](https://github.com/Texera/texera/assets/143021053/c1fad40d-1642-4eaa-9285-1a7b4d40a3cf)